validation tests complete
